refactor: remove getter and setter and move printCarState to PrintUtils
